requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-environment
django-filter
djangorestframework
filelock
gunicorn
Markdown
MouseInfo
packaging
pbr
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
stevedore
typing_extensions
tzdata
virtualenv
virtualenv-clone
virtualenvwrapper
whitenoise